Use generic channel handshake implementation for Cosmos
