+++ crates/cosmos/cosmos-chain-components/src/impls/channel/mod.rs
@@ -1,4 +1,3 @@
 pub mod channel_handshake_message;
-pub mod channel_handshake_message2;
 pub mod channel_handshake_payload;
 pub mod init_channel_options;
